# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Usuários",
        "description": "Documentação da API de Usuários",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário e gerar token JWT",
                "operationId": "350acd962ad07b73a5a36d9e556f8c9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "cre1@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Pass*a1b37"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Encerrar sessão do usuário",
                "operationId": "268291865a88e2d25df352b407f7ed78",
                "responses": {
                    "200": {
                        "description": "Logout bem-sucedido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Atualizar token JWT",
                "operationId": "58a2eaf0a23fa75d67833384a7121113",
                "responses": {
                    "200": {
                        "description": "Token atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/otp/generate": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Gerar um código OTP",
                "operationId": "6f28b164b4cbfbc759fb90a7fb5a0d11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_or_email",
                                    "type_generate"
                                ],
                                "properties": {
                                    "phone_or_email": {
                                        "type": "string",
                                        "example": "cre1@gmail.com"
                                    },
                                    "type_generate": {
                                        "type": "string",
                                        "example": "CHANGE_PASSWORD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código OTP gerado com sucesso"
                    }
                }
            }
        },
        "/api/otp/validate": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Validar um código OTP",
                "operationId": "193f827ffd1d619cc2735a96edfd22d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_or_email",
                                    "otp_code"
                                ],
                                "properties": {
                                    "phone_or_email": {
                                        "type": "string",
                                        "example": "cre1@gmail.com"
                                    },
                                    "otp_code": {
                                        "type": "string",
                                        "example": "959120"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código OTP validado com sucesso"
                    },
                    "400": {
                        "description": "Código OTP inválido ou expirado"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar todos os usuários",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Criar um novo usuário",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "johndoe@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Obter um usuário pelo ID",
                "operationId": "6f51bd5fa2dddcd25e44cfa53dcbd10a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar um usuário",
                "operationId": "070e37fc9aff78ae73860751bc990ed3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "janedoe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deletar um usuário",
                "operationId": "0b4133ed462528a5e648b88ead11e929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/change-password": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Alterar senha do usuário",
                "operationId": "e3efe22d03ce113f7ecdd8406f2fabaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "newPassword"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Pass*a1b37"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Endpoints para gerenciamento da authenticação"
        },
        {
            "name": "Usuários",
            "description": "Endpoints para gerenciamento de usuários"
        },
        {
            "name": "OTP",
            "description": "OTP"
        }
    ],
    "security": {
        "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
        }
    }
}